Big cleanup and move to new details page
