+++ immuscreen/src/app/[elementType]/[elementID]/[tab]/_IcreTabs/types.ts
@@ -1,4 +1,4 @@
-import { CellQueryValue } from "../celllineage/types"
+import { CellQueryValue } from "app/celllineage/types"
 
 
 export type ICRE_Data = { accession: string, rdhs: string, celltypes: CellQueryValue[], coordinates: { chromosome: string, start: number, end: number, } }
