+++ immuscreen/src/app/[elementType]/[elementID]/[tab]/page.tsx
@@ -7,6 +7,8 @@ import { useElementMetadata, useElementMetadataReturn } from "common/hooks/useEl
 import { GenomicElementType, isValidGeneTab, isValidIcreTab, isValidSnpTab, isValidTab } from "types/globalTypes"
 import SnpEQTLs from "./_SnpTabs/SnpEQTLs"
 import GeneEQTLs from "./_GeneTabs/GeneEQTLs"
+import GeneExpression from "./_GeneTabs/GeneExpression"
+import IcreActivity from "./_IcreTabs/IcreActivity"
 
 /**
  * @todo
@@ -33,7 +35,7 @@ export default function DetailsPage({
     return <CircularProgress />
   }
 
-  if (!elementMetadata.coordinates){
+  if (!elementMetadata?.coordinates){
     return <Typography>Issue fetching data on {elementID}</Typography>
   }
 
@@ -75,7 +77,7 @@ export default function DetailsPage({
       switch (tab) {
         case ("eQTLs"): return <GeneEQTLs name={geneData.name} id={geneData.id} />
         case ("linked"): return <p>Viewing {tab} for {elementID} in {elementType} Portal</p>
-        case ("expression"): return <p>Viewing {tab} for {elementID} in {elementType} Portal</p>
+        case ("expression"): return <GeneExpression name={geneData.name} />
       }
     }
 
@@ -88,7 +90,7 @@ export default function DetailsPage({
 
       switch (tab) {
         case ("linked"): return <p>Viewing {tab} for {elementID} in {elementType} Portal</p>
-        case ("activity"): return <p>Viewing {tab} for {elementID} in {elementType} Portal</p>
+        case ("activity"): return <IcreActivity accession={icreData.accession}/>
       }
     }
   }
