--- immuscreen/src/app/gene/types.ts
@@ -1,6 +1,0 @@
-export type PointMetaData = {
-    cellType: string;
-    name: string;
-    class: string;
-    value: number;
-}
